backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
pymongo
python-jose
passlib[bcrypt]
python-dotenv
email-validator
